spice: Split branch operations by responsibility
Move branch-name generation, branch lookup, branch mutation,
bulk branch loading, and graph-list helpers out of one shared
branch.go file and into responsibility-named files in internal/spice.
The split keeps request and result types near the operations that use
them while preserving the package API and Service behavior.

This is a package-local organization change. It also corrects nearby
comment typos in the moved code without changing branch state,
lookup, cleanup, rename, forget, or graph traversal behavior.
