+++ internal/spice/branch_name.go
@@ -1,0 +1,44 @@
+package spice
+
+import (
+	"strings"
+	"unicode"
+
+	"go.abhg.dev/gs/internal/must"
+)
+
+// GenerateBranchName generates a branch name from a commit message subject.
+//
+// The branch name is generated by converting the subject to lowercase,
+// replacing spaces with hyphens, and removing all non-alphanumeric characters.
+//
+// If the subject has more than `lengthLimit` characters,
+// it is truncated to that length at word boundaries.
+func GenerateBranchName(subject string, lengthLimit int) string {
+	words := strings.FieldsFunc(strings.ToLower(subject), func(r rune) bool {
+		return !unicode.IsLetter(r) && !unicode.IsNumber(r)
+	})
+	must.NotBeEmptyf(words, "subject must not be empty")
+
+	var name strings.Builder
+	for _, w := range words {
+		newLen := name.Len() + len(w)
+		needHyphen := name.Len() > 0
+		if needHyphen {
+			newLen++
+		}
+
+		if newLen > lengthLimit {
+			break
+		}
+
+		if needHyphen {
+			name.WriteByte('-')
+		}
+		for _, r := range w {
+			name.WriteRune(unicode.ToLower(r))
+		}
+	}
+
+	return name.String()
+}
